Clear save updated
added button to options screen in the menu to enable the user to clear their save file. Changed the clear binary file method so it overwrites player data file with default values instead of deleting the file. This was done due to a bug where the scene would attempt to load the file after it had been deleted